Redwood City TrashBin Locations
Locations of TrashBins set to a var redwoodcitytrashbinsdata. Changed file from JSON to JS, so that I can assign a variable to it and use it in the logicStep9.js code. Step not necessary but something I had done. This file can be fazed out later. Plan is to just use the original JSON file.