Add files via upload
Uploading Tire Recycle Locations In the Bay Area. I am making this one a universal file for all of the cities. For the reason that there are not too many locations in the bay area which provide this service. Therefore, it would be better to include this file for every city map.